🚚 Move NotificationEventPublisher
